requirements.txt
fastapi
uvicorn
pillow
google-genai
python-dotenv
python-multipart
streamlit
requests
